requirements.txt
asgiref==3.8.1
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
folium==0.19.4
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
xyzservices==2025.1.0
